requirements.txt
annotated-types-0.7.0 
fastapi-0.115.3 
anyio-4.6.2.post1 
exceptiongroup-1.2.2 
idna-3.10 
pydantic-2.9.2 
pydantic-core-2.23.4 
sniffio-1.3.1 
starlette-0.41.0 
typing-extensions-4.12.2
click-8.1.7 
colorama-0.4.6 
h11-0.14.0 
uvicorn-0.32.0amqp-5.2.0 
billiard-4.2.1 
celery-5.4.0 
click-didyoumean-0.3.1 
click-plugins-1.1.1 
click-repl-0.3.0 
amqp-2.6.1 
kombu-4.2.0 
vine-1.3.0
prompt-toolkit-3.0.48 
python-dateutil-2.9.0.post0 
six-1.16.0 
tzdata-2024.2 
vine-5.1.0 
wcwidth-0.2.13
aioredis-2.0.1 
async-timeout-4.0.3
PyMySQL-1.1.1 
aiomysql-0.2.0
pydantic-settings-2.6.0 
python-dotenv-1.0.1
pyyaml-6.0.2
flower-2.0.1 
humanize-4.11.0 
prometheus-client-0.21.0 
pytz-2024.2 
tornado-6.4.1
dnspython-2.7.0 
eventlet-0.37.0 
greenlet-3.1.1
redis-5.2.0
sqlalchemy-2.0.36
colorlog-6.8.2
certifi-2024.8.30 
httpcore-1.0.6 
httpx-0.27.2
dbutils-3.1.0